Move extend_tinyllama_dsus.py -> extend_tinyllama_hf_tokenizer_dsus.py
Extending the Hugging Face tokenizer is useless because the underlying SentencePiece model is not extended with new tokens - the vocabulary/tokenizer configuration files are simply extended, presumably because this is sufficient for the tokenizer to function in the context of the Hugging Face ecosystem. This does not suit out use case - wherein we need to dovetail with the wrapper of SentecePiece provided by Megatron; this uses a straight wrapper of the SentencePiece model i.e. the .model file downloaded from Hugging Face hub